add test log decorator and transfer memoized decorator in new file dop_decorators.py
