+++ src/decorators.py
@@ -8,7 +8,7 @@ def log(filename: str | None = None) -> Callable:
     """
     Пока никакого описания тут нет
     """
-    print(1)
+
     def wrapper(func: Callable) -> Callable:
         @wraps(func)
         def inner(*args: tuple, **kwargs: dict) -> Any:
@@ -31,74 +31,3 @@ def log(filename: str | None = None) -> Callable:
         return inner
 
     return wrapper
-
-
-@log(filename="mylog.txt")
-def my_function(x, y):
-    # raise TypeError("ERROR")
-    return x + y
-
-
-print(my_function(1, 2))
-
-
-def memoized(cache_count: int) -> Callable:
-    """
-    Пока никакого описания тут нет
-    """
-    cache: dict[tuple, Any] = {}
-
-    def wrapper(func: Callable) -> Callable:
-        @wraps(func)
-        def inner(*args: tuple) -> Any:
-            if args in cache:
-                return cache[args]
-            else:
-                if len(cache) >= cache_count:
-                    for key in cache.keys():
-                        del cache[key]
-                        break
-                result = func(*args)
-                cache[args] = result
-                return result
-
-        return inner
-
-    return wrapper
-
-
-@memoized(3)
-def f(x):
-    print("Calculating...")
-    return x * 10
-
-
-# print(f(1))
-# # => Calculating...
-# # 10
-# print(f(1))
-# # 10
-# print(f(42))
-# # => Calculating...
-# # 420
-# print(f(42))
-# # 420
-# print(f(1))
-# # 10
-print(f(1))
-# => Calculating...
-# 10
-print(f(1))  # Будет «вспомнено»
-# 10
-print(f(2))
-# => Calculating...
-# 20
-print(f(3))
-# => Calculating...
-# 30
-print(f(4))  # Вытеснит запомненный результат для «1»
-# => Calculating...
-# 40
-print(f(1))  # Будет перевычислено
-# => Calculating...
-# 10
